Move seq-num next to update subcommand
