+++ cmd/soroban-cli/src/commands/tx/mod.rs
@@ -1,7 +1,7 @@
 use super::global;
 
 pub mod args;
-pub mod edit;
+pub mod update;
 pub mod hash;
 pub mod help;
 pub mod new;
@@ -15,9 +15,9 @@ pub use args::Args;
 
 #[derive(Debug, clap::Subcommand)]
 pub enum Cmd {
-    /// Edit the transaction
+    /// Update the transaction
     #[command(subcommand)]
-    Edit(edit::Cmd),
+    Update(update::Cmd),
     /// Calculate the hash of a transaction envelope
     Hash(hash::Cmd),
     /// Create a new transaction
@@ -37,7 +37,7 @@ pub enum Cmd {
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
     #[error(transparent)]
-    Edit(#[from] edit::Error),
+    Update(#[from] update::Error),
     #[error(transparent)]
     Hash(#[from] hash::Error),
     #[error(transparent)]
@@ -57,7 +57,7 @@ pub enum Error {
 impl Cmd {
     pub async fn run(&self, global_args: &global::Args) -> Result<(), Error> {
         match self {
-            Cmd::Edit(cmd) => cmd.run(global_args).await?,
+            Cmd::Update(cmd) => cmd.run(global_args).await?,
             Cmd::Hash(cmd) => cmd.run(global_args)?,
             Cmd::New(cmd) => cmd.run(global_args).await?,
             Cmd::Operation(cmd) => cmd.run(global_args).await?,
